requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Environment Variables
python-decouple>=3.8

# Image Processing
Pillow>=10.0.0

# HTTP Requests (for payment gateways)
requests>=2.31.0

# Payment Gateways
stripe>=7.0.0
pycryptodome>=3.20.0

# PDF Generation
reportlab>=4.0.0

# Production Server
gunicorn>=21.0.0
whitenoise>=6.6.0

# Utilities
python-dateutil>=2.8.2
